fixing the injector

+++ integration/simulation/simulation_tester.go
@@ -22,7 +22,7 @@ func testSimulation(t *testing.T, netw network.Network, params params.SimParams)
 
 	ethClients, obscuroNodes, p2pAddrs := netw.Create(params, stats)
 
-	txInjector := NewTransactionInjector(params.NumberOfWallets, uint64(params.AvgBlockDuration), stats, ethClients, obscuroNodes)
+	txInjector := NewTransactionInjector(params.NumberOfWallets, params.AvgBlockDuration, stats, ethClients, obscuroNodes)
 
 	simulation := Simulation{
 		EthClients:       ethClients,
